+++ source/Burntime.Remaster/Logic/Location.cs
@@ -94,7 +94,7 @@ namespace Burntime.Remaster.Logic
         }
 
         protected StateLink<Player> player;
-        public Player Player
+        public Player? Player
         {
             get { return (player != null) ? player : null; }
             set { player = value; }
@@ -147,11 +147,6 @@ namespace Burntime.Remaster.Logic
             }
 
             return item;
-        }
-
-        public void RemoveItem(Item item)
-        {
-
         }
 
         // temporary
@@ -239,13 +234,7 @@ namespace Burntime.Remaster.Logic
                     if (productionState >= production.ItemDropInterval)
                     {
                         productionState -= production.ItemDropInterval;
-
-                        int i = 0;
-                        while (!Rooms[i].Items.Add(Production.Produce.Generate()))
-                        {
-                            // try next room if full
-                            i++;
-                        }
+                        StoreItem(Production.Produce.Generate());
                     }
                 }
             }
@@ -299,20 +288,22 @@ namespace Burntime.Remaster.Logic
             Items.DropAt(item, pos * Map.Mask.Resolution);
         }
 
-        // insert item in random room
-        public void StoreItemRandom(Item item)
+        public void StoreItemRandom(Item item) => StoreItem(item, randomRoom: true);
+
+        /// <summary>
+        /// Insert item into room. If none is available drop it randomly.
+        /// </summary>
+        public void StoreItem(Item item, bool randomRoom = false)
         {
-            if (Rooms.Count == 0)
+            var rooms = Rooms.Where(x => !x.Items.IsFull).ToList();
+            if (rooms.Count == 0)
             {
                 DropItemRandom(item);
                 return;
             }
 
-            Room room;
-            do
-            {
-                room = this.Rooms[Burntime.Platform.Math.Random.Next() % Rooms.Count];
-            } while (room.Items.MaxCount != ItemList.Infinite && room.Items.MaxCount < room.Items.Count);
+            int index = randomRoom ? Platform.Math.Random.Next(0, rooms.Count - 1) : 0;
+            var room = rooms[index];
 
             // fill up empty bottles
             if (room.IsWaterSource && item.Type.Full != null && Source.Reserve >= item.Type.Full.WaterValue)
@@ -341,5 +332,67 @@ namespace Burntime.Remaster.Logic
             return null;
         }
         #endregion
+
+        public void ClearItemsAfterTakeover()
+        {
+            if (container.Root is not ClassicGame root)
+                return;
+
+            foreach (var room in Rooms)
+                room.Items.Clear();
+
+            var list = new List<Item?>();
+
+            // all levels
+            if (root.World.Difficulty <= 2)
+            {
+                // random food
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "food" }, Array.Empty<string>(), 1));
+
+                // 20% chance of a rare/special
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "rare", "special" }, new string[] { "nodrop" }, 0.2f));
+
+                // 50% chance of a bottle or canteen
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "bottle" }, Array.Empty<string>(), 0.5f));
+
+                // 0-2 random items
+                list.AddRange(root.ItemTypes.GenerateClass(new string[] { "material", "useless" }, Array.Empty<string>(), 0, 2));
+            }
+
+            // level 2 and 1
+            if (root.World.Difficulty <= 1)
+            {
+                // 50% chance of a food trap
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "trap" }, Array.Empty<string>(), 0.5f));
+
+                // another 20% chance of a rare/special
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "rare", "special" }, Array.Empty<string>(), 0.2f));
+
+                // 0-2 random items
+                list.AddRange(root.ItemTypes.GenerateClass(new string[] { "material", "useless" }, Array.Empty<string>(), 0, 2));
+            }
+
+            // level 1 only
+            if (root.World.Difficulty == 0)
+            {
+                // random food
+                list.AddRange(root.ItemTypes.GenerateClass(new string[] { "food" }, Array.Empty<string>(), 1, 2));
+
+                // another 50% chance of a food trap
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "trap" }, Array.Empty<string>(), 0.5f));
+
+                // another 20% chance of a rare/special
+                list.Add(root.ItemTypes.GenerateClass(new string[] { "rare", "special", "protection_parts" }, Array.Empty<string>(), 0.2f));
+
+                // 1-2 random items
+                list.AddRange(root.ItemTypes.GenerateClass(new string[] { "material", "useless" }, Array.Empty<string>(), 1, 2));
+            }
+
+            foreach (var item in list)
+            {
+                if (item is not null)
+                    StoreItemRandom(item);
+            }
+        }
     }
 }